feat: add discord message service.🐥
